Resolve Navbar
